Two new types of representations and encapsulated
