+++ README.md
@@ -1,5 +1,5 @@
 # MetaCountdown
 
-![mETAcOUNTDOWN logo](./docs/twitter_header_photo_1.png)
+![MetaCountdown logo](./docs/twitter_header_photo_1.png)
 
 [![license](https://img.shields.io/github/license/mashape/apistatus.svg?maxAge=2592000)](https://github.com/vicdoja/MetaCountdown/blob/main/LICENSE)
