+++ tests/deap_linear_tree_test.py
@@ -11,9 +11,9 @@ from deap import base
 from deap import creator 
 from deap import tools
 
-from metacountdown.individuals import GeneralTreeIndividual
+from metacountdown.individuals import LinearTreeIndividual
 
-from metacountdown.utils import ALL_NUMBERS, DIFFICULT_INSTANCES, eval_polish
+from metacountdown.utils import ALL_NUMBERS, DIFFICULT_INSTANCES, eval_linear
 
 def init(obj=None, poss_num=None):
     if not obj:
@@ -37,7 +37,7 @@ def init(obj=None, poss_num=None):
     #toolbox.register("map", pool.map)
 
     # Attribute generator
-    toolbox.register("gen_tree", GeneralTreeIndividual.generate_individual, \
+    toolbox.register("gen_tree", LinearTreeIndividual.generate_individual, \
       	pos_num=poss_num)
 
     # Structure initializers
@@ -45,11 +45,11 @@ def init(obj=None, poss_num=None):
       	toolbox.gen_tree, n=1)
     toolbox.register("population", tools.initRepeat, list, toolbox.individual)
 
-    toolbox.register("evaluate", GeneralTreeIndividual.evaluate_individual, \
+    toolbox.register("evaluate", LinearTreeIndividual.evaluate_individual, \
       	objective=obj, pos_num=poss_num)
 
-    toolbox.register("mate", GeneralTreeIndividual.mate_individual)
-    toolbox.register("mutate", GeneralTreeIndividual.mutate_individual, \
+    toolbox.register("mate", LinearTreeIndividual.mate_individual)
+    toolbox.register("mutate", LinearTreeIndividual.mutate_individual, \
       	pos_num=poss_num, indpb=0.05)
     toolbox.register("select", tools.selTournament, tournsize=10)
 
@@ -58,7 +58,7 @@ def init(obj=None, poss_num=None):
 def main():
     random.seed(64)
     verbose = True
-    max_gen = 300
+    max_gen = 10
     pop_size = 3000
 
     for obj, poss_num in DIFFICULT_INSTANCES[:]:
@@ -71,7 +71,7 @@ def main():
         stats.register("valid", lambda pop: (numpy.array(pop) < 10**6).sum())
         stats.register("invalid", lambda pop: (numpy.array(pop) >= 10**6).sum())
         stats.register("% of valid", \
-          	lambda pop: (numpy.array(pop) < 10**6).sum()/len(pop))
+         	lambda pop: (numpy.array(pop) < 10**6).sum()/len(pop))
         
         pop, log = algorithms.eaSimple(pop, toolbox, cxpb=0.5, mutpb=0.2, \
           	ngen=max_gen, stats=stats, halloffame=hof, verbose=verbose)
@@ -81,8 +81,9 @@ def main():
         print("-"*30)
         print("The objective number is %d \nAnd the usable numbers are %s" % \
           	(obj, poss_num))
-        print(list(zip(map(lambda o: str(o[0]),hof), [(eval_polish(i[0].expr), \
-			abs(eval_polish(i[0].expr)-obj)) for i in hof])))
+        print(list(zip(map(lambda o: str(o[0]),hof), \
+            [(eval_linear(poss_num, i[0].nums, i[0].ops), \
+			abs(eval_linear(poss_num,i[0].nums,i[0].ops)-obj)) for i in hof])))
         print()
 
 if __name__ == "__main__":
